requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
tensorflow==2.12.0
numpy==1.23.5
Pillow==9.5.0
matplotlib==3.7.1
scikit-learn==1.2.2
pandas==2.0.0
werkzeug==2.2.3
reportlab==3.6.12
requests==2.26.0
opencv-python>=4.6.0
